# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample recipes API",
    "title": "Recipes API",
    "contact": {
      "name": "Luis Oropeza",
      "url": "https://www.gitlab.com/luisfi",
      "email": "luis.oropeza@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
    "/refresh": {
      "post": {
        "description": "Refresh token",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "parameters": [
          {
            "type": "string",
            "description": "token",
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Invalid token"
          },
          "404": {
            "description": "Token is not expired yet"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/signin": {
      "post": {
        "description": "Authenticate user",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signin",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Password",
            "name": "password",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/jwtOutput"
            }
          },
          "401": {
            "description": "Invalid username or password"
          },
          "404": {
            "description": "Invalid data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "jwtOutput": {
      "type": "object",
      "properties": {
        "expires": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Expires"
        },
        "token": {
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-name": "JWTOutput",
      "x-go-package": "go-microservices-example/handlers"
    }
  }
}